+++ packages/lwc-plugin-pretty-histogram-series/tests/graphics/radius-and-width.js
@@ -15,7 +15,7 @@ function runTestCase(container) {
 	const chart = (window.chart = LightweightCharts.createChart(container, {
 		layout: { attributionLogo: false },
 	}));
-	const series = chart.addCustomSeries(new LwcPlugin.PrettyHistogramSeries(), {
+	const series = LwcPlugin.createPrettyHistogramSeries(chart, {
 		color: '#2962FF',
 		widthPercent: 90,
 		radius: 12,