+++ packages/lwc-plugin-stacked-bars-series/src/example/example.ts
@@ -1,5 +1,5 @@
 import { CustomSeriesWhitespaceData, Time, createChart } from 'lightweight-charts';
-import { StackedBarsSeries } from '../stacked-bars-series';
+import { createStackedBarsSeries } from '../stacked-bars-series';
 import { StackedBarsColumnWidthMode } from '../options';
 import { StackedBarsData } from '../data';
 import { multipleBarData } from './sample-data';
@@ -29,7 +29,7 @@ const chart = ((window as unknown as any).chart = createChart('chart', {
 	},
 }));
 
-const series = chart.addCustomSeries(new StackedBarsSeries(), {
+const series = createStackedBarsSeries(chart, {
 	color: 'black', // for the price line
 });
 series.setData(demoData());